jesd204: remove link index from ops callback
This seems to cause more confusion that is helpful. After some usage of the
current form of the framework, it looks like it would be more useful to
just know the JESD204 link IDs as they are provided in the framework, and
let the framework care about the index in the array of link objects.

This wasn't clear in the beginning, but it is more clear now [that is more
confusing than useful].

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 587d682e8133f98f6dcbaf65d07ed2b99dd1beca)