jesd204: move state definitions into the dt-bindings include
If we want to define which states should be sync'ed for multi-SOM setups,
we will need to define which states required synchronization during system
boot-up and also JESD204 link init (or re-init).

To do this, we need to move all the JESD204 states in the dt-bindings
location to have them accessible to the device-trees, so that we can add
properties that act on this behavior.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 395f5f09743dbcf379564ae170a0d1d499d44e2c)