iio: adrv9002: api: Integrate version 20.9.2
This patch fixes the errors and warnings introduced when compiling the
navassa API in the kernel. The fixes here are only meant to the API.
Further fixes to adapt the adrv9002 driver with this version will come in a
future patch.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>
(cherry picked from commit 938a467aa8442b7ffebb59921317fb958f2a1208)