iio: adrv9002: Only register the needed clocks
This patch refactors on the clocks are handled. In first place, a
pointer is added in the channeld structure to provide a fastpath to
access the interface clock.

On the other hand, if we are in rx2tx2 mode, the driver only needs to
export 2 clocks since RX2/TX2 will use the same clock as RX1/TX1.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>
(cherry picked from commit 42a1a924859f547cf004d366836e4cf5428cdf15)