iio: adrv9002: Add some delays when changing ensm
For operations that need to be in a specific ensm to succeed, we need to
add some delays to make sure that the device actually goes into that
state before proceeding with the operation. This problem was visible
when changing the LO frequency for a port with the enable mode set to
pin.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>
(cherry picked from commit 604054187fda8d33931f4acb4a9fe01452dd28b2)