iio: beamformer: adar1000: Change gain to DAC code
The gain value is not linear. The ADAR1000 is not ment to be used on its
own so there is no direct conversion to dB for gain value. This patch
changes the value for gain to DAC codes.

Signed-off-by: Mircea Caprioru <mircea.caprioru@analog.com>
(cherry picked from commit 2deb583c866620d13d8aed2284067bd7b169d58b)