iio: adrv9002: refactor error message for dts properties
This patch is a minor improvement on the helper macro to get 'u32' dts
properties. For mandatory properties, in case of error, we were just
assuming that the property is missing. However, 'EINVAL', is not the only
error that 'of_property_read_u32()' can return. Hence we just print the
error code and print that we failed to get a mandatory property.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>
(cherry picked from commit 1812d78eb65aff93d3610de37dad6e9bdd9ea5d7)