net: ethernet: adi: adin1110: NAPI for ADIN1110
A failed attempt of adding NAPI support for ADIN1110 driver.
Not possible as retrieving the ethernet frames requires us
to sleep in the poll function, thing that is not allowed.

One way to circumvent this is to do that in a separate
thread or to do that in the bottom-half of the used IRQ and
store the ethernet frames in a multiframe skb/fifo. Only after
that to send them up the stack (during polling).

Won't improve the driver's performance by much, maybe just
adjust the ethernet frames flow through the stack.

Signed-off-by: Alexandru Tachici <alexandru.tachici@analog.com>